+++ SpedeSpelitBaseline2024/buttons.h
@@ -5,21 +5,14 @@
 #include <avr/interrupt.h>
 
 
-const byte firstPin = 2; // First PinChangeInterrupt on D-bus
-const byte lastPin =  5; // Last PinChangeInterrupt on D-bus
+const byte buttonPins[] = {2,3,4,5,6};
+const byte numButtons = sizeof(buttonPins) / sizeof(buttonPins[0]);
 
-/* 
-  initButtonsAndButtonInterrupts subroutine is called from Setup() function
-  during the initialization of Speden Spelit. This function does the following:
-  1) Initializes 4 button pins for the game = Arduino pins 2,3,4,5
-  2) Initializes 1 button pin for starting the game = Aruino pin 6
-  3) Enables PinChangeInterrupt on D-bus in a way that interrupt
-     is generated whenever some of pins 2,3,4,5,6 is connected to LOW state
+extern volatile unsigned long lastDebounceTime[numButtons];
+const unsigned long debounceDelay = 150;
 
-*/
 void initButtonsAndButtonInterrupts(void);
 
 
-// Intoduce PCINT2_vect Interrupt SeRvice (ISR) function for Pin Change Interrupt.
-ISR(PCINT2_vect); 
-#endif;
+ISR(PCINT2_vect);
+#endif
