--- SpedeSpelitBaseline2024/buttons/buttons1_copy_20241128015708.ino
@@ -1,12 +1,0 @@
-#include "buttons.h"
-
-
-void setup(){
-
-  Serial.begin(9600);
-  initButtonsAndButtonInterrupts();
-}
-
-void loop() { 
-
-}
